requirements.txt
asgiref==3.8.1
cloudinary==1.42.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
jsonfield==3.1.0
oauthlib==3.2.2
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
requests-oauthlib==2.0.0
sqlparse==0.5.3
